+++ CHANGELOG.md
@@ -6,6 +6,8 @@
 ### Added
 - bourbon (http://bourbon.io) for development
 - add attribute sundayHighlight
+- add attribute maxYear
+- add attribute minYear
 
 ### Fixed
 - reset view to calendar after click cancel on select year option